+++ src/css/deprecated-css-custom-property-syntax.ts
@@ -1,0 +1,111 @@
+/**
+ * @license
+ * Copyright (c) 2017 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
+import {Document, ParsedCssDocument, Severity} from 'polymer-analyzer';
+import * as shady from 'shady-css-parser';
+
+import {CssRule} from '../css/rule';
+import {registry} from '../registry';
+import {FixableWarning} from '../warning';
+
+import stripIndent = require('strip-indent');
+
+
+class DeprecatedCustomPropertySyntax extends CssRule {
+  code = 'deprecated-css-custom-property-syntax';
+  description = stripIndent(`
+      Warns when using deprecated css syntax around CSS Custom Properties.
+  `).trim();
+
+  async checkDocument(parsedDocument: ParsedCssDocument, _document: Document) {
+    const warnings: FixableWarning[] = [];
+
+    for (const node of parsedDocument) {
+      this.addAtApplyWarnings(node, parsedDocument, warnings);
+      this.addVarSyntaxWarnings(node, parsedDocument, warnings);
+    }
+
+    return warnings;
+  }
+
+  // Convert `@apply(--foo);` to `@apply foo;`
+  private addAtApplyWarnings(
+      node: shady.Node, parsedDocument: ParsedCssDocument,
+      warnings: FixableWarning[]) {
+    if (node.type === shady.nodeType.atRule && node.name === 'apply') {
+      if (node.parametersRange && node.parameters.startsWith('(') &&
+          node.parameters.endsWith(')')) {
+        const w = new FixableWarning({
+          code: 'at-apply-with-parens',
+          parsedDocument,
+          severity: Severity.ERROR,
+          sourceRange:
+              parsedDocument.sourceRangeForShadyRange(node.parametersRange),
+          message:
+              '@apply with parentheses is deprecated. Prefer: @apply --foo;'
+        });
+        w.fix = [
+          {
+            range: parsedDocument.sourceRangeForShadyRange({
+              start: node.parametersRange.start,
+              end: node.parametersRange.start + 1
+            }),
+            replacementText: ' '
+          },
+          {
+            range: parsedDocument.sourceRangeForShadyRange({
+              start: node.parametersRange.end - 1,
+              end: node.parametersRange.end
+            }),
+            replacementText: ''
+          }
+        ];
+        warnings.push(w);
+      }
+    }
+  }
+
+  // Convert `var(--foo, --bar)` to `var(--foo, var(--bar))`
+  private addVarSyntaxWarnings(
+      node: shady.Node, parsedDocument: ParsedCssDocument,
+      warnings: FixableWarning[]): any {
+    if (node.type === 'expression') {
+      const match = node.text.match(
+          /var\s*\(\s*--[a-zA-Z0-9_-]+\s*,\s*(--[a-zA-Z0-9_-]+)\s*\)/);
+      if (match) {
+        const offsetOfVarInsideExpression = match.index!;
+        const offsetOfSecondCustomPropWithinVar =
+            match[0].match(/--[a-zA-Z0-9_-]+\s*\)$/)!.index!;
+        const secondCustomProp = match[1];
+        const newText = `var(${secondCustomProp})`;
+        const start = node.range.start + offsetOfVarInsideExpression +
+            offsetOfSecondCustomPropWithinVar;
+        const end = start + secondCustomProp.length;
+        const sourceRange =
+            parsedDocument.sourceRangeForShadyRange({start, end});
+        const warning = new FixableWarning({
+          code: 'invalid-second-arg-to-var-expression',
+          severity: Severity.WARNING, parsedDocument, sourceRange,
+          message:
+              'When the second argument to a var() expression is another ' +
+              'custom property, it must also be wrapped in a var().'
+        });
+        warning.fix = [{range: sourceRange, replacementText: newText}];
+        warnings.push(warning);
+      }
+    }
+  }
+}
+
+registry.register(new DeprecatedCustomPropertySyntax());