+++ src/linter.ts
@@ -15,6 +15,7 @@
 import './collections';
 
 import {Analyzer, Document, ParsedDocument, Severity, Warning, WarningCarryingException} from 'polymer-analyzer';
+import {FixableWarning} from 'warning';
 
 import {Rule} from './rule';
 
@@ -39,7 +40,7 @@ export class Linter {
    * Given an array of filenames, lint the files and return an array of all
    * warnings produced evaluating the linter rules.
    */
-  public async lint(files: string[]): Promise<Warning[]> {
+  public async lint(files: string[]): Promise<FixableWarning[]> {
     const {documents, warnings} = await this._analyzeAll(files);
     for (const document of documents) {
       warnings.push(...document.getWarnings());
@@ -47,7 +48,7 @@ export class Linter {
     return warnings.concat(...await this._lintDocuments(documents));
   }
 
-  public async lintPackage(): Promise<Warning[]> {
+  public async lintPackage(): Promise<FixableWarning[]> {
     const pckage = await this._analyzer.analyzePackage();
     const warnings = pckage.getWarnings();
     warnings.push(
@@ -56,8 +57,13 @@ export class Linter {
   }
 
   private async _lintDocuments(documents: Iterable<Document>) {
-    const warnings: Warning[] = [];
+    const warnings: FixableWarning[] = [];
     for (const document of documents) {
+      if (document.isInline) {
+        // We lint the toplevel documents. If a rule wants to check inline
+        // documents, it can. getFeatures makes that pretty easy.
+        continue;
+      }
       for (const rule of this._rules) {
         try {
           warnings.push(...await rule.check(document));
@@ -78,6 +84,7 @@ export class Linter {
     const analysis = await this._analyzer.analyze(files);
     const documents = [];
     const warnings = [];
+
 
     for (const file of files) {
       const result = analysis.getDocument(this._analyzer.resolveUrl(file));