+++ src/rules.ts
@@ -12,6 +12,8 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
+import './css/deprecated-css-custom-property-syntax';
+
 import './html/dom-module-name-or-is';
 import './html/move-style-into-template';
 import './html/undefined-elements';